Remove WalletConnect v1 completely
Remove WalletConnect v1 completely from the codebase. We are removing
`@web3-react/walletconnect-connector` library that was used to implement
WalletConnect v1, but we are keeping one of it's dependencies which were
`@walletconnect/types` in `1.7.8` version needed to implement WalletConnect v2.

Besides that we are now renaming all occurences of WalletConnectV2 to just
WalletConnect. We also removes the possibility to choose WalletConnectv1 in the
`SelectWallet` modal.
