+++ package.json
@@ -45,12 +45,12 @@
     "@types/react-router-dom": "^5.3.2",
     "@walletconnect/ethereum-provider": "^2.8.6",
     "@walletconnect/modal": "^2.5.9",
+    "@walletconnect/types": "1.7.8",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
     "@web3-react/ledger-connector": "^6.1.9",
     "@web3-react/trezor-connector": "^6.1.9",
     "@web3-react/types": "^6.0.7",
-    "@web3-react/walletconnect-connector": "^6.2.4",
     "@web3-react/walletlink-connector": "^6.2.13",
     "axios": "^0.24.0",
     "bitcoin-address-validation": "^2.2.1",
